[#18] MemberDTO -> Member 이름 변경
