--- src/main/java/com/icia/funding/controller/TestController.java
@@ -1,17 +1,0 @@
-package com.icia.funding.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class TestController {
-
-	
-	@GetMapping("/healthy")
-	@ResponseBody
-	public String healthy() {
-		return "200 ok";
-	}
-	
-}
