+++ src/main/java/io/github/mygoodsupporter/domain/OAuthToken.java
@@ -1,10 +1,6 @@
-package com.icia.funding.dto;
+package io.github.mygoodsupporter.domain;
 
 import lombok.Data;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
 
 @Data
 public class OAuthToken {
