+++ src/main/java/io/github/mygoodsupporter/security/MemberDetails.java
@@ -1,13 +1,11 @@
-package com.icia.funding.security;
+package io.github.mygoodsupporter.security;
 
 
-import com.icia.funding.dto.MemberDTO;
-import lombok.Data;
+import io.github.mygoodsupporter.domain.Member;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.ArrayList;
@@ -25,9 +23,9 @@ public class MemberDetails implements UserDetails {
     private String email;
 
 
-    public MemberDetails(MemberDTO member) {
+    public MemberDetails(Member member) {
         this.id = member.getM_id();
-        this.password = "";
+        this.password = member.getM_password();
         this.enabled = member.getEnabled();
         this.authorities = member.getAuthorities();
         this.email = member.getM_email();
