+++ src/main/resources/mapper/MemberMapper.xml
@@ -1,25 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="membermapper">
+<mapper namespace="MemberMapper">
 	<!-- 회원가입 -->
-	<insert id="memberjoin" parameterType="memberdto">
+	<insert id="insertMember" parameterType="Member">
 		INSERT INTO MEMBER(M_ID,M_PASSWORD,M_NAME,M_PHONE,M_EMAIL,M_TYPE, ENABLED) VALUES(#{m_id},#{m_password},#{m_name},#{m_phone},#{m_email},#{m_type},1)
 	</insert>
 
-			<!-- 아이디 체크 -->
-	<select id="idcheck" parameterType="String" resultType="String">
-	SELECT M_ID FROM MEMBER WHERE M_ID=#{m_id}
+	<select id="getMemberById" parameterType="String" resultType="Member">
+		SELECT * FROM MEMBER WHERE M_ID = #{m_id}
 	</select>
-
-			<!-- 로그인 -->
-	<select id="memberlogin" parameterType="memberdto" resultType="String">
-	SELECT M_ID FROM MEMBER WHERE M_ID=#{m_id} AND M_PASSWORD=#{m_password}
-	</select>
-
-	<select id="getMemberById" parameterType="String" resultType="memberdto">
-		SELECT * FROM MEMBER WHERE M_ID=#{m_id}
-	</select>
-
 
 </mapper>
