+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -27,7 +27,7 @@
 	</bean>
 
 
-	<context:component-scan base-package="com.icia.funding" use-default-filters="false">
+	<context:component-scan base-package="io.github.mygoodsupporter" use-default-filters="false">
 		<context:include-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
 	</context:component-scan>
