+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -43,7 +43,7 @@
 	
 	<tx:annotation-driven transaction-manager="transactionManager"/>
 
-	<context:component-scan base-package="com.icia.funding">
+	<context:component-scan base-package="io.github.mygoodsupporter">
 		<context:exclude-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
 	</context:component-scan>
 </beans>
