+++ src/main/webapp/WEB-INF/spring/security-context.xml
@@ -7,9 +7,10 @@
 
     <security:http auto-config="true" use-expressions="true">
         <security:intercept-url pattern="/" access="permitAll"/>
-        <security:intercept-url pattern="/order" access="hasRole('ROLE_MEMBER')"/>
+        <security:intercept-url pattern="/users/member" access="hasRole('ROLE_MEMBER')"/>
         <security:intercept-url pattern="/users/admin" access="hasRole('ROLE_ADMIN')"/>
         <security:form-login login-page="/login" login-processing-url="/login" username-parameter="m_id" password-parameter="m_password"/>
+
         <security:logout logout-url="/logout" logout-success-url="/"/>
         <security:csrf disabled="true"></security:csrf>
     </security:http>
@@ -22,7 +23,7 @@
         </security:authentication-provider>
     </security:authentication-manager>
 
-    <bean id="MemberDetailsService" class="com.icia.funding.security.MemberDetailsService" />
+    <bean id="MemberDetailsService" class="io.github.mygoodsupporter.security.MemberDetailsService" />
     <bean id="passwordEncoder" class="org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder"/>
 
 </beans>
