+++ src/test/java/com/github/mygoodsupporter/dao/MemberDAOTest.java
@@ -1,35 +1,36 @@
-package com.icia.funding.service;
+package com.github.mygoodsupporter.dao;
 
-import com.icia.funding.dto.MemberDTO;
+import io.github.mygoodsupporter.dao.MemberDAO;
+import io.github.mygoodsupporter.domain.Member;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.transaction.annotation.Transactional;
 
 import static org.junit.Assert.*;
 
+
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration({"file:src/main/webapp/WEB-INF/spring/root-context.xml", "file:../../../../../../main/webapp/WEB-INF/spring/security-context.xml"})
-//@Transactional
-public class MemberServiceTest {
+@ContextConfiguration("file:src/main/webapp/WEB-INF/spring/root-context.xml")
+@Transactional
+public class MemberDAOTest {
 
     @Autowired
-    MemberService memberService;
+    MemberDAO memberDAO;
 
     @Test
-    public void joinTest() {
-
-        MemberDTO member = new MemberDTO();
-        member.setM_id("abba123");
+    public void memberJoin() {
+        Member member = new Member();
+        member.setM_id("aa123");
         member.setM_password("1234");
         member.setM_name("이름");
         member.setM_phone("010-2944-4442");
         member.setM_email("bbbb1234@email.com");
         member.setM_type("ADMIN");
 
-        int result = memberService.memberJoin(member);
-
+        int result = memberDAO.insertMember(member);
         assertEquals(1, result);
     }
 }
