+++ src/test/java/com/github/mygoodsupporter/service/MemberServiceTest.java
@@ -1,37 +1,36 @@
-package com.icia.funding.dao;
+package com.github.mygoodsupporter.service;
 
-import com.icia.funding.dto.MemberDTO;
-import org.junit.Assert;
+import io.github.mygoodsupporter.domain.Member;
+import io.github.mygoodsupporter.service.MemberService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.transaction.annotation.Transactional;
 
 import static org.junit.Assert.*;
 
-
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration("file:src/main/webapp/WEB-INF/spring/root-context.xml")
-@Transactional
-public class MemberDAOTest {
+@ContextConfiguration({"file:src/main/webapp/WEB-INF/spring/root-context.xml", "file:../../../../../../main/webapp/WEB-INF/spring/security-context.xml"})
+//@Transactional
+public class MemberServiceTest {
 
     @Autowired
-    MemberDAO memberDAO;
+    MemberService memberService;
 
     @Test
-    public void memberJoin() {
-        MemberDTO member = new MemberDTO();
-        member.setM_id("aa123");
+    public void joinTest() {
+
+        Member member = new Member();
+        member.setM_id("abba123");
         member.setM_password("1234");
         member.setM_name("이름");
         member.setM_phone("010-2944-4442");
         member.setM_email("bbbb1234@email.com");
         member.setM_type("ADMIN");
 
-        int result = memberDAO.memberJoin(member);
+        int result = memberService.join(member);
+
         assertEquals(1, result);
     }
 }
