+++ src/test/resources/log4j.xml
@@ -11,7 +11,7 @@
 	</appender>
 	
 	<!-- Application Loggers -->
-	<logger name="com.icia.funding">
+	<logger name="com.github.mygoodsupporter">
 		<level value="info" />
 	</logger>
